Big Refactoring

+++ MyLocationsIOS/Model/AppRepository.swift
@@ -10,19 +10,28 @@ import CoreData
 
 struct AppRepository {
     
-    fileprivate var locations: [Location]
+    fileprivate var locations: [Location]?
     fileprivate let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
     
     mutating func fetchLocations() {
         do {
             self.locations = try context.fetch(Location.fetchRequest())
         } catch {
-            print("Wir konnten leider die Kontakte nicht fetchen.")
+            print (error.localizedDescription)
         }
-        notifyObservers()
     }
     
-    func getLocations() -> [Location] {
+    func fetchLocations() -> [Location]? {
+        do {
+            let locations = try context.fetch(Location.fetchRequest())
+        } catch {
+            print(error.localizedDescription)
+        }
+        notifyObservers()
+        return locations
+    }
+    
+    func getLocations() -> [Location]? {
         return self.locations
     }
     
@@ -31,8 +40,10 @@ struct AppRepository {
     }
     
     mutating func updateLocation(at index: Int, with location: Location) {
-        self.locations[index] = location
-        notifyObservers()
+        if var loc = self.locations {
+            loc[index] = location
+            notifyObservers()
+        }
     }
     
     mutating func addLocation(name: String, description: String, w3w: String, longitude: Float, latitude: Float) {
@@ -57,22 +68,28 @@ struct AppRepository {
     }
     
     mutating func deleteLocation(_ location: Location) {
-        for index in 0 ... self.locations.count {
-            if self.locations[index].name == location.name {
-                deleteLocation(at: index)
+        if let locs = self.locations {
+            for index in 0 ... locs.count {
+                if locs[index].name == location.name {
+                    deleteLocation(at: index)
+                }
             }
         }
     }
     
     mutating func deleteLocation(at index: Int) {
-        let locationToDelete = self.locations[index]
-        do {
-            self.context.delete(locationToDelete)
-            try self.context.save()
-            self.locations.remove(at: index)
-        } catch {
-            print(error.localizedDescription)
+        
+        if let locs = self.locations {
+            let locationToDelete = locs[index]
+            do {
+                self.context.delete(locationToDelete)
+                try self.context.save()
+            } catch {
+                print(error.localizedDescription)
+            }
+            notifyObservers()
         }
+        self.locations?.remove(at: index)
         notifyObservers()
     }
